form updated to  DecimalField from IntegerField
Form now accept decimal data and doesn't clear the field when submitting from the edit link in table

+++ app/forms.py
@@ -6,7 +6,7 @@ Created on Mon Jul 23 20:51:50 2018
 """
 
 from flask_wtf import FlaskForm
-from wtforms import StringField, PasswordField, BooleanField, SubmitField, SelectField, IntegerField, DateField
+from wtforms import StringField, PasswordField, BooleanField, SubmitField, SelectField, IntegerField, DateField, DecimalField
 from wtforms.validators import ValidationError, DataRequired, Email, EqualTo
 from app.models import User
 from datetime import date
@@ -24,8 +24,8 @@ class MiniForm(FlaskForm):
      date = StringField('Date', default = date.today)
      project = SelectField('Project', choices=[('project1', 'project1'), ('project2','project2')])
      task = SelectField('Task', choices=[('task1', 'task1'), ('task2', 'task2'), ('task3', 'task3')])
-     predicted_hrs = IntegerField('Predicted Hours')
-     actual_hrs = IntegerField('Actual Hours')
+     predicted_hrs = DecimalField('Predicted Hours')
+     actual_hrs = DecimalField('Actual Hours')
      submit = SubmitField('Submit')
      
 class RegistrationForm(FlaskForm):
